+++ common/src/main/java/com/gitlab/enums/PickupPointFeatures.java
@@ -1,0 +1,16 @@
+package com.gitlab.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+@AllArgsConstructor
+@Getter
+public enum PickupPointFeatures {
+
+    TRY_ON_CLOTHES("Примерка одежды"),
+    TRY_ON_SHOES("Примерка обуви"),
+    DELIVERY_FOR_BUSINESSES("Доставка для юридических лиц"),
+    PARTIAL_ORDER_REDEMPTION("Частичный выкуп заказа"),
+    PRODUCT_RETURNS("Возврат товаров");
+
+    private final String pickupPointFeatureRussianTranslation;
+}