+++ gazon-service/src/main/resources/db/changelog/v-1.17/02-fill-payments-table.sql
@@ -1,0 +1,14 @@
+INSERT INTO payments (bank_card_id, payment_status, create_date_time, order_id, sum, user_id)
+VALUES (1, 'NOT_PAID', '2023-01-21 11:01', 1, 1500, 1);
+INSERT INTO payments (bank_card_id, payment_status, create_date_time, order_id, sum, user_id)
+VALUES (2, 'PAID', '2023-02-22 12:02', 1, 2500, 2);
+INSERT INTO payments (bank_card_id, payment_status, create_date_time, order_id, sum, user_id)
+VALUES (3, 'OVERDUE', '2023-03-23 13:03', 1, 3500, 3);
+
+
+
+
+
+
+
+