+++ gazon-service/src/test/java/com/gitlab/controller/PickupPointRestControllerIT.java
@@ -1,8 +1,8 @@
 package com.gitlab.controller;
 
 import com.gitlab.dto.PickupPointDto;
+import com.gitlab.enums.PickupPointFeatures;
 import com.gitlab.mapper.PickupPointMapper;
-import com.gitlab.model.PickupPoint;
 import com.gitlab.service.PickupPointService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -140,7 +140,7 @@ class PickupPointRestControllerIT extends AbstractIntegrationTest {
         pickupPointDto.setAddress("TestAddress");
         pickupPointDto.setDirections("TestDirections");
         pickupPointDto.setShelfLifeDays((byte) 16);
-        pickupPointDto.setPickupPointFeatures(Set.of(PickupPoint.PickupPointFeatures.values()));
+        pickupPointDto.setPickupPointFeatures(Set.of(PickupPointFeatures.values()));
 
         return pickupPointDto;
     }