+++ gazon-service/src/test/java/com/gitlab/mapper/OrderMapperTest.java
@@ -21,7 +21,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class OrderMapperTest extends AbstractIntegrationTest {
 
-
     @Autowired
     private OrderMapper mapper;